ci: rename actions

+++ install-linux.sh
@@ -1,4 +1,12 @@
+#!/bin/bash
+
 rm -f bin/IsoMaker
-cmake cmake -B cmake/build
-make -C cmake/build
-mv cmake/build/IsoMaker bin
+cmake -B cmake/build -S cmake
+cmake --build cmake/build
+
+if [ -f cmake/build/IsoMaker ]; then
+    mv cmake/build/IsoMaker bin
+else
+    echo "Compilation failed: IsoMaker binary not found!"
+    exit 1
+fi
